#23 move lightning network providers to their own folder
